add a solution for track code = -1(mute music) change trackCode type from float to int32

+++ Assets/BCIPlugin/src/Services/Trial.cs
@@ -90,7 +90,7 @@ public class Trial
         }
         else if (cmd[0] == "SetTrack")
         {
-            var trackCode = Convert.ToSingle(cmd[1]);
+            var trackCode = Convert.ToInt32(cmd[1]);
             BCIPlugin.GetComponent<AAEOnlineParadigm>().SetTrack(trackCode);
         }
     }
